refactor(simulation): extract machine-file parsing into machine.py
Move the machine-description parser/validator and its model dataclasses
(SimChannel, Scenario) out of engine.py into a new machine.py, behind a single
parse_machine() -> ParsedMachine entry point that engine.__init__ now calls.
DEFAULT_SCENARIO and the event-shape table move with them; SimReading (a runtime
result) stays in engine.py. engine.py 664 -> 390 lines. Add direct parser tests.
